Mrc pixel sizes (#52)
* Revert "provides option to have shape visually centered if box size is even"

This reverts commit bf359a9c2a268b6f157a3598e6cb8ae164a07ca3.

* mrc includes voxel sizes (lost upon a prior commit)
variables put into consistent order
tests updated accordingly
soft edge requires int

+++ src/ttmask/soft_edge.py
@@ -2,7 +2,7 @@ import numpy as np
 from scipy.ndimage import distance_transform_edt
 
 
-def add_soft_edge(mask: np.ndarray, width: float) -> np.ndarray:
+def add_soft_edge(mask: np.ndarray, width: int) -> np.ndarray:
     distance_from_edge = distance_transform_edt(mask == 0)
     boundary_pixels = (distance_from_edge <= width) & (distance_from_edge != 0)
     normalised_distance_from_edge = (distance_from_edge[boundary_pixels] / width) * np.pi
